Introduce auto releases for windows/macos/linux
This pipeline introduces not only `intuit/auto` for automatic release
versioning, but also includes builds for the cli for windows, macos, and
linux.

Due to a problem with gh-actions caching that I never nailed down, I
fell back to using the runner VMs directly. Since our setup is pretty
minimal anyway, and caching actually works with the VMs, I just went
with it. I did wrestle with getting caching to work for windows, which
ultimately lead me to this bug report:

https://github.com/actions/cache/issues/591#issuecomment-845132253

That represents at least 6 hours of my life I won't get back, and is
a good example why I like circleci and gitlabci so much more than
gh-actions...

But gh-actions frustration aside, this should generate a release with a
reasonable changelog and binaries uploaded to the release for each os.

DBAAS-6841