Add create database command
This adds the new `create` verb along with a command for `database` and
its related flags. I went ahead and bound all of the flags with viper so
they can be provided via the environment var or config file just in case
someone wanted to do that.

I also didn't give all of the options a short name due to conflicts, but
I tried to prioritize the ones I thought might be used the most.

While testing the command, I noticed that our old examples are now
rejected by SNOW, so I pinged Eric and Mohsin to see what's up there.
Depending on their response, we may need to change the examples in the
API and the defaults here in the command.

But assuming that biz rule gets worked out, you would be able to create
a database instance with reasonable defaults with...

```
skysqlcli create database --name dmayo-cli --provider "Amazon AWS"
--region="ca-central-1"
```

DBAAS-7917