Extract flag defaults into constants.go
This was a suggestion by John Halbert to make updating default values
easier later. It uses more code, but it does centralize all the defaults
into one spot without the noise of the rest of the flag definitions
distracting from them.

It is likely that we'll be thinking about the defaults independently
from the rest of the flag definitions, so from a human standpoint
pulling them out and grouping the defaults all together will help us
more easily spot deviations in the defaults chosen for each command.

DBAAS-6841