Avoid viper bug regarding flag name collisions
This issue on github explains the problem:
https://github.com/spf13/viper/issues/233#issuecomment-386791444

The first collision we noticed was between the update and create
commands for databases, where the update command flag for `--name`
overwrote the flag for create - which resulted in it not being passed
along for the create command.

By waiting to do the binding until the prerun, we don't have to worry
about multiple commands colliding with each other since only the command
that is run will execute the prerun - which results in the config file
and env var resolution behaving like we expected.

Note also that with this change the config file will now follow the flag
names (using kebab case), and only the env vars will use snake case.

DBAAS-6841