Use kebab case for api-key flag in config file
To be clear, no one should be putting a plaintext private api key into
the config file as that is a _bad idea_ from a security perspective.

But, for consistency sake, that flag uses kebab case not snake case, so
the viper lookup should follow along. The environment variable will
still work as intended, as we have the `SetEnvKeyReplacer` already set
up to swap out dashes for low dashes for the environment variables.

DBAAS-6841