Fix auto version included in go build of cli
What I have there still works locally, but apparently it isn't working
in the github actions workflow.

I saw someone else having the same trouble over in the github issues for
intuit/auto, so I'm following the pattern they used for a workaround to
hopefully address the issue.

https://github.com/intuit/auto/issues/2062#issue-979861914

I don't have a good way to test this, other than to merge and find out,
without doing a lot of changes to the pipeline. Hopefully this works as
is, but if it doesn't then I'll go back and find a way to test so that
everyone isn't pestered by me working on the pipeline.

DBAAS-6841