Install additional plugins for `intuit/auto` in build pipeline
Yikes - so we have that install for auto happen twice in the pipeline,
and in that build job we didn't include the plugins. I really with
gh-actions gave us more to work with for deduping this kind of stuff.

At some point I guess we'll have to start writing custom actions.

I also noticed in the cobra docs that if you populate the version prop
then it adds a `--version` flag to the root command which is nice.

DBAAS-6841