Determine version info at start of pipeline
I pulled out the auto usage from the build matrix because it was doing
the same thing in each of those three jobs, and due to them being
different environments it ran into trouble with windows.

I figure since it's the same information, just calculate it once and pass
the value along to the build jobs.

I also went ahead and added a conditional on the release job, that way
we can just add the branch we're working on to the list of branches at
the top and test out the build without accidentally making a release.

DBAAS-6841