Generate cli docs as md, yaml, rst, and man pages
With the hope of helping out the docs team, this commit introduces a
new make target (`make docs`) which executes a script (`docs/main.go`)
that will walk the command tree for the skysqlcli and generate
documentation based on the `--help` for every command.

Since I imagine that different people will want that information in
different forms, I went ahead and generated it in each format that cobra
had implemented. That is...

* Markdown
* YAML
* ReSTructured text
* Linux manuals

This will hopefully make it a bit easier for someone walking up to the
cli for the first time to be able to browse around the docs before
committing to downloading the binaries and giving it a spin.

DBAAS-6841