Install golang for release job in pipeline
Since I'm using that script to generate the docs automatically, I need
to have golang installed and configured for our private repos for it to
be able to walk through the command. We _could_ make a binary for the
docs gen, but it's only ever used in the pipeline so I figure a go run
is fine.

DBAAS-6841