Include autogen docs in staged content for release commit
My make target just generates the docs, but doesn't actually add them to
the staging area in git. So when `intuit/auto` goes to make the release
commit, the check they have to verify the working directory is clean
runs afoul of those docs - which is good since I intended for those to
be included!

This change should add them to the staging area so they are included in
the commit - though I don't have an easy way to test that other than
merging this PR, so we'll see if that was the only issue.

DBAAS-6841