Add `skysqlcli get credentials` command for default creds
Now that the API has an endpoint to expose the default credentials for a
service, this adds a command to leverage it. To be clear, these
credentials are only meant for temporary, initial setup of a service and
should be removed once the user has set up their own users in the
database themselves.

Note that this PR is using a pre-release of the go sdk until this PR
goes in:
https://github.com/mariadb-corporation/skysql-api-go/pull/24

DBAAS-8521