Introduce new `skysqlcli get allowlist-status` command
This new command works with the new endpoint in the api which exposes
whether an allowlist is still updating or if it is done updating and is
enforcing.

Since this change brings in the new sdk, it also includes the update to
all of the allowlist endpoints now that they are under the `/security/`
path. The change is transparent behind the sdk, so nothing really to see
here in this PR for that.

DBAAS-8823